Update tests to use junit4 annotations
Every test method is annotated with a @Test annotation. The naming is
irrelevant. Test classes don't need to be an instance of TestCase anymore.

Thus I eliminated the TestBase class and instead made use of junit4 rules, in
particular the TemporaryFolder rule. I also moved all other helper methods in
this class although this is not nice and could be improved.

Some tests need XMLUnit.setIgnoreWhitespace(true) which is now present only in
two setUp methods. Some test classes might fail now if run individually and
the call has not been made by another test class. However it would be nice, if
this global state could be eliminated somehow. This must be investigated in
the XMLUnit library.